fixed brickDaoTests, added DataAccessException subclass on service layer, added this exceptions to brickService class

+++ src/main/java/cz/muni/fi/PA165/service/BrickServiceImpl.java
@@ -16,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import org.springframework.dao.DataAccessException;
 
 /**
  *
@@ -33,16 +34,25 @@ public class BrickServiceImpl implements BrickService {
     }
 
     public void create(BrickDto brickDto) {
+        if(brickDto == null) {
+            throw new DataAccessExceptionService("created brick cannot be null");
+        }
         Brick brickEntity = BrickConversion.convertToEntity(brickDto);
         brickDao.create(brickEntity);
     }
 
     public void update(BrickDto brickDto) {
+        if(brickDto == null) {
+            throw new DataAccessExceptionService("updated brick cannot be null");
+        }
         Brick brickEntity = BrickConversion.convertToEntity(brickDto);
         brickDao.update(brickEntity);
     }
 
     public void delete(Long id) {
+        if(id == null) {
+            throw new DataAccessExceptionService("cannot delete brick with null id");
+        }
         brickDao.delete(id);
     }
 
@@ -56,11 +66,17 @@ public class BrickServiceImpl implements BrickService {
     }
 
     public BrickDto findById(Long id) {
+        if(id == null) {
+            throw new DataAccessExceptionService("cannot find brick with null id");
+        }
         Brick entity = brickDao.findById(id);
         return BrickConversion.convertToDto(entity);
     }
 
     public List<BrickDto> findByColor(Color color) {
+        if(color == null) {
+            throw new DataAccessExceptionService("cannot find brick with null color");
+        }
         List<Brick> brickEntitites = brickDao.findByColor(color);
         List<BrickDto> brickDtoList = new ArrayList<BrickDto>();
         for (Brick brick : brickEntitites) {
@@ -70,6 +86,9 @@ public class BrickServiceImpl implements BrickService {
     }
 
     public List<BrickDto> findByName(String name) {
+        if(name == null) {
+            throw new DataAccessExceptionService("cannot brick with null name");
+        }
         List<Brick> brickEntitites = brickDao.findByName(name);
         List<BrickDto> brickDtoList = new ArrayList<BrickDto>();
         for (Brick brick : brickEntitites) {
